backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.0
celery[redis]==5.3.4
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
aiohttp==3.9.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.15.0
opencv-python==4.8.1.78
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
pylint==3.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0
freezegun==1.2.2